Backend/requirements.txt
# Criar um novo requirements.txt simplificado

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
requests==2.31.0
aiofiles==0.8.0
PyJWT==2.8.0
httpx==0.25.2
jinja2==3.1.2
python-docx==1.1.0
reportlab==4.0.7
pandas==2.1.4
matplotlib==3.8.2
#"@ | Out-File -FilePath requirements.txt -Encoding UTF8